# doc/openapi.yaml
openapi: 3.0.0
info:
  title: SiteOps Web API
  description: API specification for the SiteOps web application
  version: 1.0.0
  contact:
    name: SiteOps Team
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.siteops.local
    description: Production server
paths:
...
  /api/project/{projectID}/assigned-people:
    get:
      summary: Get people assigned to a project
      operationId: getAssignedPeopleByProject
      tags:
        - Project
      description: Returns all people assigned to the specified project.
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: integer
          description: Project ID
          example: 1
      responses:
        '200':
          description: Assigned people retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  people:
                    type: array
                    items:
                      $ref: '#/components/schemas/Person'
        '400':
          description: Invalid project ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

# Activity endpoints
...
components:
  schemas:
...
    Person:
      type: object
      description: Person record in the system
      properties:
        personID:
          type: integer
          description: Unique person identifier
          example: 1
        name:
          type: string
          description: Person name
          example: Nimal Perera
        position:
          type: string
          description: Job title or position
          example: Site Engineer
        yearsOfExperience:
          type: integer
          nullable: true
          description: Years of experience
          example: 8
        NIC:
          type: string
          description: National identity card number
          example: 991234567V
      required:
        - personID
        - name
        - position
        - NIC
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or signup endpoint